Fix Decimal

+++ report-service/app/events.py
@@ -26,7 +26,7 @@ class AccountCreatedEvent(BaseEventWithUUID):
     iban: str = Field(..., min_length=5, max_length=34)
     bic: str = Field(..., min_length=7, max_length=11)
     name: str = Field(..., min_length=1, max_length=255)
-    balance: int = Field(0, ge=0)
+    balance: Decimal = Field(0, ge=0, max_digits=12, decimal_places=2)
     currency: Currency = Field(..., min_length=3, max_length=3)
